# internal/docs/generated/swagger.yaml
basePath: /api
definitions:
  apicommon.APIErrorBody:
    properties:
      code:
        description: 错误码
        type: string
      details:
        type: object
      message:
        description: 错误描述
        type: string
    type: object
  apicommon.APIErrorResponse:
    properties:
      error:
        allOf:
        - $ref: '#/definitions/apicommon.APIErrorBody'
        description: 错误体
    type: object
  v1.V1ErrorEnvelope:
    properties:
      error:
        description: 错误体
        type: string
    type: object
...
  v2.V2ChargingCurveAPIResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/v2.V2ChargingCurveResponse'
        description: 响应数据
      meta:
        allOf:
        - $ref: '#/definitions/v2.V2Meta'
        description: 响应元信息
    type: object
  v2.V2ChargingCurveResponse:
    properties:
      min_sessions:
        type: integer
      samples:
        description: 样本
        items:
          $ref: '#/definitions/v2.V2ChargingCurveSample'
        type: array
    type: object
  v2.V2ChargingCurveSample:
    properties:
      avg_voltage:
        type: number
      battery_level:
        description: 电量百分比
        type: integer
      max_power:
        description: 该 SoC 桶内会话功率上限 (kW)
        type: number
      median_power:
        type: number
      min_power:
        description: 该 SoC 桶内会话功率下限 (kW)
        type: number
      p25_power:
        type: number
      p75_power:
        type: number
      session_count:
        description: 充电会话数
        type: integer
    type: object
...
  v2.V2Meta:
    properties:
      applied_filters:
        additionalProperties:
          type: string
        description: |-
          AppliedFilters echoes back the query-string filters the server actually
          honored (audit §5.2 / spec §5.2). Helps clients debug "why did I get this
          payload?" — they can compare what they sent vs. what the server applied
          (defaults, normalized casing, ignored unknown keys, etc.). Always
          non-nil when populated; serialized as a JSON object.
        type: object
      car_id:
        description: 车辆 ID
        example: 1
        type: integer
      compare:
        description: 对比模式
        enum:
        - none
        - previous_period
        type: string
      end:
        description: 结束时间
        format: date-time
        type: string
      generated_at:
        format: date-time
        type: string
      period:
        description: 聚合周期
        enum:
        - day
        - week
        - month
        - quarter
        - year
        - custom
        - lifetime
        type: string
      start:
        description: 起始时间
        format: date-time
        type: string
      timezone:
        description: 时区
        example: Asia/Shanghai
        type: string
      unit:
        allOf:
        - $ref: '#/definitions/v2.V2Unit'
        description: 单位
    type: object
...
  v2.V2Unit:
    properties:
      consumption:
        example: Wh/km
        type: string
      currency:
        example: CNY
        type: string
      distance:
        description: 距离 (km)
        example: km
        type: string
      duration:
        description: 时长 (秒)
        example: seconds
        type: string
      elevation:
        example: m
        type: string
      energy:
        example: kWh
        type: string
      power:
        example: kW
        type: string
      speed:
        example: km/h
        type: string
      temperature:
        example: C
        type: string
    type: object
...
info:
  contact: {}
  description: TeslaMate 数据查询与分析 API。提供 V1（原始数据）和 V2（聚合分析）两套接口。
  title: TeslaMateApi
  version: "2.0"
paths:
...
  /v1/cars/{CarID}/tire-pressure:
    get:
      description: 最近一次 TPMS 读数 + 时间窗内极值 + 按天历史。每日历史返回 *_min / *_max（用于发现慢漏气）和 *_avg（已弃用，下个
        minor 删除）。日历桶按 apicommon.AppUsersTimezone (TZ 环境变量) 切分，避免非 UTC 用户跨午夜 off-by-one。气压单位遵循
        settings.unit_of_pressure（默认 bar；psi 由 conv.BarToPsi 换算）。
      parameters:
      - description: 车辆 ID
        in: path
        name: CarID
        required: true
        type: integer
      - description: 历史窗口天数（默认 30，最大 365）
        in: query
        name: window_days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.V1ErrorEnvelope'
      summary: 车辆胎压（最近读数 + 窗口）
      tags:
      - v1
...
  /v2/cars/{CarID}/analytics/charging/curve:
    get:
      description: 返回该车直流充电曲线的聚合样本（按电量分桶：会话数、功率分布 min/p25/median/p75/max）。默认覆盖整个生命周期；显式传入
        start/end/period 时按窗口过滤。
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 聚合周期
        enum:
        - day
        - week
        - month
        - quarter
        - year
        - custom
        in: query
        name: period
        type: string
      - description: 起始时间（RFC3339）
        in: query
        name: start
        type: string
      - description: 结束时间（RFC3339）
        in: query
        name: end
        type: string
      - description: IANA 时区
        in: query
        name: timezone
        type: string
      - description: 每个电量分桶的最少会话数（默认 5）
        in: query
        name: min_sessions
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.V2ChargingCurveAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apicommon.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apicommon.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apicommon.APIErrorResponse'
      summary: V2 直流充电曲线（聚合）
      tags:
      - v2
...
swagger: "2.0"
...